Module Dockerfile_linux.Zypper

Rules for Zypper-based distributions such as OpenSUSE

val update : Dockerfile.t

update will run zypper update non-interactively.

val install : ('a, unit, string, Dockerfile.t) Stdlib.format4 -> 'a

install fmt will zypper update && zypper install the packages specified by the fmt format string.

val add_user : ?⁠uid:int -> ?⁠gid:int -> ?⁠sudo:bool -> string -> Dockerfile.t

add_user username will install a new user with name username and a locked password. If sudo is true then root access with no password will also be configured. The default value for sudo is false.

val dev_packages : ?⁠extra:string -> unit -> Dockerfile.t

dev_packages ?extra () will install the base development tools and sudo, passwd and git. Extra packages may also be optionally supplied via extra.

val install_system_ocaml : Dockerfile.t

Install the system OCaml packages via zypper